WIP - stdlib

+++ c/common/basic/code/core/symbols.c
@@ -23,7 +23,6 @@
 /*
 #include "../library/include/string.h"
 */
-#include <stdint.h>
 #include "string.h"
 #include "stdio.h"
 
@@ -51,7 +50,7 @@ void initialise_symbol_table( void )
 {
 
 	unsigned int alter;
-	alter = 16 - ( ((unsigned int) (uintptr_t) top ) % 16);
+	alter = 16 - ( ((unsigned int) (uint32_t) top ) % 16);
 
 	// initialise the heap (base & top) pointers to be just above any stored program (align to 16 byte boundary)
 	heap_base         = top + alter;
