+++ c/common/basic/code/language/next.c
@@ -19,8 +19,6 @@
 // 8 < - - - - - - - - - - 8 < - - - - - - - - - - 8 < - - - - - - - - - - 8 < - - - - - - - - - - 8 < - - - - - - - - - - 8 < - - - - - - - - - - 8 < - - - - - - - - - -
 
 
-#include <stdint.h>
-
 // include the library
 #include "stdlib.h"
 #include "stdio.h"
@@ -71,7 +69,7 @@ void _next( void )
 	
 	popByte();
 	basetype             = popByte();
-	source_ptr           = (unsigned char *)(uintptr_t) popInteger();
+	source_ptr           = (unsigned char *)(uint32_t) popInteger();
 	switch ( basetype )
 	{
 		case DATATYPE_INTEGER :
@@ -137,7 +135,7 @@ void _next( void )
 				pushReal( st_entry_finish.rval );
 				break;
 		}
-		pushInteger( (unsigned int) (uintptr_t) source_ptr );
+		pushInteger( (unsigned int) (uint32_t) source_ptr );
 		pushByte   ( basetype             );
 		pushByte   ( DATATYPE_FOR         );
 		runtime_ptr        = (unsigned char *) ( source_ptr );
